backend/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
gunicorn==21.2.0
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
zipp==3.17.0
